requirements.txt
# Face Recognition Dependencies
deepface>=0.0.93
tensorflow>=2.12.0,<2.16.0  # Compatible with deepface
tf-keras-vis

# Core Flask Dependencies
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
sqlalchemy==2.0.30
Flask-CORS==6.0.1
cryptography>=45.0.0

# Computer Vision & Machine Learning
opencv-contrib-python==4.11.0.86
numpy==1.26.4
scikit-learn==1.7.0
matplotlib==3.10.3
Pillow==11.3.0

# QR Code Generation/Reading
qrcode==8.2
pyzbar==0.1.9

# Face Recognition (Fallback)
# dlib  # Commented out due to CMake dependency issues
face_recognition  # Requires dlib; ensure build tools are available
# face-recognition  # Commented out due to dlib dependency

# MediaPipe for Face Detection & Landmarks
mediapipe==0.10.21

# Optional: Deep Learning Acceleration
# tensorflow-gpu  # Uncomment if you have NVIDIA GPU with CUDA support

# Development & Utilities
Werkzeug==3.1.3

# Additional dependencies
imutils==0.5.4
scipy==1.11.4
scikit-image==0.21.0
reportlab==4.0.8 